% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bb_pwer}
\alias{bb_pwer}
\title{Power Analysis Results}
\format{
\subsection{\code{bb_pwer}}{

A data frame with 76 rows and 2 columns:
\describe{
\item{N}{Sample size}
\item{Pwr}{Rejection rate}
}
}
}
\usage{
bb_pwer
}
\description{
Results from power analysis with a 0.5 slope of interest and a 0.1 correlation
between predictors.
}
\details{
Simulation framework modified from:
https://stats.stackexchange.com/questions/1866/how-to-simulate-a-custom-power-analysis-of-an-lm-model-using-r
Framework presented on stackexchange thread cites Bolker (2008) <ISBN: 9780691125220>
}
\keyword{datasets}
