% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{punc_pwr}
\alias{punc_pwr}
\title{Puncture: Statistical Power}
\format{
\subsection{\code{punc_pwr}}{

A list with two elements:
\describe{
\item{results}{Full results from simulation}
\item{summary}{Summary of results from simulation}
}
}
}
\usage{
punc_pwr
}
\description{
Results from puncture with a 0.5 slope of interest.
}
\keyword{datasets}
