% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{punc_pwr_het}
\alias{punc_pwr_het}
\title{Puncture: Statistical Power with Heteroskedasticity}
\format{
\subsection{\code{punc_pwr_het}}{

A list with two elements:
\describe{
\item{results}{Full results from simulation}
\item{summary}{Summary of results from simulation}
}
}
}
\usage{
punc_pwr_het
}
\description{
Results from puncture with a 0.5 slope of interest and heteroskedasticity.
}
\keyword{datasets}
