% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{punc_type1}
\alias{punc_type1}
\title{Puncture: Type 1 Error}
\format{
\subsection{\code{punc_type1}}{

A list with two elements:
\describe{
\item{results}{Full results from simulation}
\item{summary}{Summary of results from simulation}
}
}
}
\usage{
punc_type1
}
\description{
Results from puncture type 1 error rate analysis
}
\keyword{datasets}
